requirements.txt
playwright==1.42.0
google-generativeai==0.5.4
beautifulsoup4==4.12.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
lxml==5.1.0
